Move chai-chain-utils to utils folder
