--- src/transformers/util/index.js
@@ -1,14 +1,0 @@
-const createCall = require('./createCall');
-const chainContains = require('./chainContains');
-const getNodeBeforeMemberExpression = require('./getNodeBeforeMemberExpression');
-const updateExpect = require('./updateExpect');
-const createCallChain = require('./createCallChain');
-
-
-module.exports = {
-    createCall,
-    chainContains,
-    getNodeBeforeMemberExpression,
-    updateExpect,
-    createCallChain,
-};
